Better approximate totem's line

+++ src/game/game.rs
@@ -536,23 +536,23 @@ impl Game {
     fn spawn_totems(&mut self) {
         let mut rng = rand::thread_rng();
         self.board.board.retain(|t| t.is_zord());
-        let (x_c, y_c) = (BASE_BOARD_SIZE / 2, BASE_BOARD_SIZE / 2);
+        let (x_c, y_c) = ((BASE_BOARD_SIZE / 2) as f32, (BASE_BOARD_SIZE / 2) as f32);
         loop {
             let (x_t, y_t) = (
-                rng.gen_range(0..BASE_BOARD_SIZE),
-                rng.gen_range(0..BASE_BOARD_SIZE),
+                rng.gen_range(0..BASE_BOARD_SIZE) as f32,
+                rng.gen_range(0..BASE_BOARD_SIZE) as f32,
             );
-            if (x_t - x_c).abs() == 0 || (y_t - y_c).abs() == 0 {
+            if (x_t - x_c).abs() as i16 == 0 || (y_t - y_c).abs() as i16 == 0 {
                 continue;
             }
 
             let m = ((y_c - y_t).abs()) / ((x_c - x_t).abs());
             let q = y_t - (m * x_t);
-            let f = |x: i16| x * m + q;
+            let f = |x: f32| x * m + q;
             let diff = (x_c - x_t).abs();
 
-            let t1 = (x_c - diff, f(x_c - diff));
-            let t2 = (x_c + diff, f(x_c + diff));
+            let t1 = ((x_c - diff) as i16, f(x_c - diff) as i16);
+            let t2 = ((x_c + diff) as i16, f(x_c + diff) as i16);
 
             let is_far_enough =
                 (t1.0 - t2.0).abs().max((t1.1 - t2.1).abs()) > (TOTEM_AURA * 2) as i16;
